+++ config/webpack.common.js
@@ -4,7 +4,7 @@ const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
-const { title, description, author } = require('../siteMeta');
+const { title, description, author } = require('../src/siteMeta');
 
 const paths = require('./paths');
 
@@ -80,6 +80,13 @@ module.exports = {
       '@': path.resolve(__dirname, '../src'),
       '@assets': path.resolve(__dirname, '../src/assets'),
       '@components': path.resolve(__dirname, '../src/components'),
+      '@services': path.resolve(__dirname, '../src/services'),
+      '@store': path.resolve(__dirname, '../src/store'),
+      '@utils': path.resolve(__dirname, '../src/utils'),
+      '@constants': path.resolve(__dirname, '../src/utils/constants'),
+      '@helpers': path.resolve(__dirname, '../src/utils/helpers'),
+      '@hooks': path.resolve(__dirname, '../src/utils/hooks'),
+      '@views': path.resolve(__dirname, '../src/views'),
     },
     extensions: ['.js', '.jsx'],
   },
