+++ src/components/content/index.js
@@ -1,0 +1,6 @@
+import ErrorMessage from './ErrorMessage/ErrorMessage';
+import HtmlContent from './HtmlContent/HtmlContent';
+import Logo from './Logo/Logo';
+import Status from './Status/Status';
+
+export { ErrorMessage, HtmlContent, Logo, Status };
