+++ src/components/controls/index.js
@@ -1,0 +1,6 @@
+import Button from './Button/Button';
+import Checkbox from './Checkbox/Checkbox';
+import TextField from './TextField/TextField';
+import Toggle from './Toggle/Toggle';
+
+export { Button, Checkbox, TextField, Toggle };
