+++ src/components/layout/index.js
@@ -1,0 +1,6 @@
+import Container from './Container/Container';
+import Header from './Header/Header';
+import ScrollToTop from './ScrollToTop/ScrollToTop';
+import Wrapper from './Wrapper/Wrapper';
+
+export { Container, Header, ScrollToTop, Wrapper };
