--- src/components/pages/Position/index.js
@@ -1,1 +1,0 @@
-export { default } from './Position';
