+++ src/views/Job/index.js
@@ -1,0 +1,1 @@
+export { default as JobPage } from './JobPage/JobPage';
